fix: Show Header Logo immediately for non-home pages
On the home page, Header logo is hidden due to the Hero section
containing larger version of it. As user scrolls down, the Header logo
eventually appears. This effect is present in all pages, which is not
desirable.

Ensure Header logo is hidden only on a home page - not all pages.